fix review findings: thread safety, stale closure, god object decomposition
CopyOnWriteArrayList for concurrent job tracking in BaseMonitoringOrchestrator.
Remove as-any cast in App.tsx with proper ModalName typing.
Extract useWeather, useAIChat, useSessionHistory hooks from 913-line
ModalManager (reduced to 558 lines).
